Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-gcm-import-key-decrypt-additional-data-expected.txt

Test decrypting using AES-GCM with an imported 128bit key and additional data

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToASCIIString(plainText) is expectedPlainText
PASS successfullyParsed is true

TEST COMPLETE
